Move type
